+++ packages/at_client/lib/src/client/remote_secondary.dart
@@ -246,7 +246,7 @@ class RemoteSecondary implements Secondary {
   }
 
   void addStreamData(List<int> data) {
-    atLookUp.connection!.getSocket().add(data);
+    atLookUp.connection!.add(data);
   }
 
   /// Generates digest using from verb response and [secret] and performs a CRAM authentication to